Revert "refactor(server-platform): consolidate server-database into platform"
This reverts commit 36604d819f70544f084b8198434ff5e8de61b735.
