+++ packages/server-database/oxlint.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from "oxlint";
+
+import baseConfig from "../../oxlint.config.ts";
+
+export default defineConfig({
+  extends: [baseConfig],
+  plugins: ["node"],
+  rules: {},
+  overrides: [],
+});
