+++ packages/server-database/src/index.test.ts
@@ -1,0 +1,5 @@
+import { expect, test } from "bun:test";
+
+test("1 == 1", () => {
+  expect(1).toEqual(1);
+});
