+++ packages/server-database/src/index.ts
@@ -1,0 +1,2 @@
+export { createPrismaClient } from "./client.ts";
+export { PrismaClient } from "./generated/client.ts";
