+++ packages/server-database/tsconfig.json
@@ -5,5 +5,5 @@
     "outDir": "./dist"
   },
   "exclude": ["node_modules"],
-  "include": ["src"]
+  "include": ["src", "prisma.config.ts"]
 }
