+++ packages/server-platform/package.json
@@ -6,25 +6,15 @@
     ".": "./src/index.ts"
   },
   "scripts": {
-    "db:generate": "prisma generate",
-    "db:migrate:deploy": "prisma migrate deploy",
-    "db:migrate:dev": "prisma migrate dev",
-    "db:migrate:reset": "prisma migrate reset",
-    "db:studio": "prisma studio",
     "lint": "oxlint",
     "test": "bun test --randomize"
   },
   "dependencies": {
-    "@prisma/adapter-pg": "catalog:",
-    "@prisma/client": "catalog:",
-    "pg": "catalog:",
     "pino": "catalog:",
     "pino-pretty": "catalog:",
     "zod": "catalog:"
   },
   "devDependencies": {
-    "@repo/typescript-config": "workspace:*",
-    "@types/pg": "catalog:",
-    "prisma": "catalog:"
+    "@repo/typescript-config": "workspace:*"
   }
 }
