+++ packages/server-platform/src/index.ts
@@ -1,4 +1,2 @@
 export * from "./config.ts";
 export * from "./logger.ts";
-export { createPrismaClient } from "./prisma-client.ts";
-export { PrismaClient } from "./generated/client.ts";
